requirement.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
names==0.3.0
python-dateutil==2.8.1
pytz==2021.1
randomtimestamp==1.0.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
